pyproject.toml
[project]
name = "image-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "openpyxl>=3.1.5",
    "pillow>=12.1.1",
    "selenium>=4.41.0",
    "webdriver-manager>=4.0.2",
]
